Move files to package.
