+++ packages/e2e-test-utils/README.md
@@ -1,0 +1,3 @@
+# E2E test utils
+
+A set of e2e utils for web stories tests.
