+++ packages/e2e-test-utils/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@web-stories-wp/e2e-test-utils",
+  "private": true,
+  "engines": {
+    "node": ">= 15",
+    "npm": "~7"
+  },
+  "type": "module",
+  "exports": {
+    "default": "./src/index.js"
+  },
+  "main": "./src/index.js",
+  "dependencies": {
+    "@wordpress/e2e-test-utils": "^4.16.1",
+    "uuid": "^8.3.2"
+  }
+}
