+++ packages/e2e-test-utils/src/addTextElement.js
