+++ packages/e2e-test-utils/src/deleteMedia.js
