+++ packages/e2e-test-utils/src/uploadMedia.js
@@ -16,7 +16,7 @@
 /**
  * Internal dependencies
  */
-import { uploadFile } from './';
+import { uploadFile } from './index';
 
 const MODAL = '.media-modal';
