+++ packages/e2e-test-utils/src/visitDashboard.js
